feat: Adiciona protecao aos códigos do consoleController

+++ bd-projeto-guiado-III/gameStore/src/controller/consolesController.js
@@ -28,8 +28,21 @@ const findAllConsoles = async (req, res) => {
 
 const findConsoleById = async (req, res) => {
   try {
-    const findConsole = await ConsolesModel.findById(req.params.id);
-    res.status(200).json(findConsole);
+    const authHeader = req.get("authorization");
+    if (!authHeader) {
+      return res.status(401).send("Você esqueceu de passar as informações de autorização");
+    }
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token,
+      SECRET,
+      async function(erro){
+        if (erro) {
+          return res.status(403).send("Acesso não autorizado.");
+        }
+
+        const findConsole = await ConsolesModel.findById(req.params.id);
+        res.status(200).json(findConsole);
+      });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: error.message });
@@ -38,30 +51,44 @@ const findConsoleById = async (req, res) => {
 
 const addNewConsole = async (req, res) => {
   try {
-    const {
-      name,
-      developer,
-      releaseDate,
-      display,
-      storageCapacities,
-      numberOfPlayers,
-      available,
-      description,
-    } = req.body;
-    const newConsole = new ConsolesModel({
-      name,
-      developer,
-      releaseDate,
-      display,
-      storageCapacities,
-      numberOfPlayers,
-      available,
-      description,
-    });
-
-    const savedConsole = await newConsole.save();
-
-    res.status(201).json({ message: "New console successfully added", savedConsole });
+    const authHeader = req.get("authorization");
+    if (!authHeader) {
+      return res.status(401).send("Você esqueceu de passar as informações de autorização");
+    }
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token,
+      SECRET,
+      async function(erro){
+        if (erro) {
+          return res.status(403).send("Acesso não autorizado.");
+        }
+        
+        const {
+          name,
+          developer,
+          releaseDate,
+          display,
+          storageCapacities,
+          numberOfPlayers,
+          available,
+          description,
+        } = req.body;
+        const newConsole = new ConsolesModel({
+          name,
+          developer,
+          releaseDate,
+          display,
+          storageCapacities,
+          numberOfPlayers,
+          available,
+          description,
+        });
+    
+        const savedConsole = await newConsole.save();
+    
+        res.status(201).json({ message: "New console successfully added", savedConsole });
+      });
+    
   } catch (error) {
     console.error(error);
     res.status(500).json(error.message);
@@ -70,28 +97,42 @@ const addNewConsole = async (req, res) => {
 
 const updateConsole = async (req, res) => {
   try {
-    const {
-      name,
-      developer,
-      releaseDate,
-      display,
-      storageCapacities,
-      numberOfPlayers,
-      available,
-      description,
-    } = req.body;
-    const updateConsole = await ConsolesModel.findByIdAndUpdate(req.params.id, {
-      name,
-      developer,
-      releaseDate,
-      display,
-      storageCapacities,
-      numberOfPlayers,
-      available,
-      description,
-    });
+    const authHeader = req.get("authorization");
+    if (!authHeader) {
+      return res.status(401).send("Você esqueceu de passar as informações de autorização");
+    }
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token,
+      SECRET,
+      async function(erro){
+        if (erro) {
+          return res.status(403).send("Acesso não autorizado.");
+        }
 
-    res.status(200).json({ message: "Console successfully updated", updateConsole });
+        const {
+          name,
+          developer,
+          releaseDate,
+          display,
+          storageCapacities,
+          numberOfPlayers,
+          available,
+          description,
+        } = req.body;
+        const updateConsole = await ConsolesModel.findByIdAndUpdate(req.params.id, {
+          name,
+          developer,
+          releaseDate,
+          display,
+          storageCapacities,
+          numberOfPlayers,
+          available,
+          description,
+        });
+    
+        res.status(200).json({ message: "Console successfully updated", updateConsole });
+      });
+    
   } catch {
     console.error(error);
     res.status(500).json({ message: error.message });
@@ -100,10 +141,23 @@ const updateConsole = async (req, res) => {
 
 const deleteConsole = async (req, res) => {
   try {
-    const { id } = req.params;
-    const deleteConsole = await ConsolesModel.findByIdAndDelete(id);
-    const message = `Console with id ${deleteConsole.name}, was successfully deleted`;
-    res.status(200).json({ message });
+    const authHeader = req.get("authorization");
+    if(!authHeader){
+      return res.status(401).send("Você esqueceu de passar as informações de autorização");
+    }
+    const token = authHeader.split(" ")[1];
+    jwt.verify(token,
+      SECRET,
+      async function(erro){
+        if (erro) {
+          return res.status(403).send("Acesso não autorizado.");
+        }
+      
+        const { id } = req.params;
+        const deleteConsole = await ConsolesModel.findByIdAndDelete(id);
+        const message = `Console with id ${deleteConsole.name}, was successfully deleted`;
+        res.status(200).json({ message });
+    });
   } catch (error){
     console.error(error);
     res.status(500).json({ message: error.message });
